Split metrics code into files
