+++ pkg/plugin/metrics/node.go
@@ -1,0 +1,69 @@
+package metrics
+
+import (
+	"github.com/neondatabase/autoscaling/pkg/plugin/state"
+	"github.com/neondatabase/autoscaling/pkg/util"
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+type nodeMetrics struct {
+	// InheritedLabels are the labels on the node that are directly used as part of the metrics
+	InheritedLabels []string
+
+	cpu *prometheus.GaugeVec
+	mem *prometheus.GaugeVec
+}
+
+func buildNodeMetrics(labels nodeLabeling, reg prometheus.Registerer) nodeMetrics {
+	finalMetricLabels := []string{"node"}
+	finalMetricLabels = append(finalMetricLabels, labels.metricLabelNames...)
+	finalMetricLabels = append(finalMetricLabels, "field")
+
+	return nodeMetrics{
+		InheritedLabels: labels.k8sLabelNames,
+
+		cpu: util.RegisterMetric(reg, prometheus.NewGaugeVec(
+			prometheus.GaugeOpts{
+				Name: "autoscaling_plugin_node_cpu_resources_current",
+				Help: "Current amount of CPU for 'state.NodeResources' fields",
+			},
+			finalMetricLabels,
+		)),
+		mem: util.RegisterMetric(reg, prometheus.NewGaugeVec(
+			prometheus.GaugeOpts{
+				Name: "autoscaling_plugin_node_mem_resources_current",
+				Help: "Current amount of memory (in bytes) for 'state.NodeResources' fields",
+			},
+			finalMetricLabels,
+		)),
+	}
+}
+
+func (m *nodeMetrics) Update(node *state.Node) {
+	commonLabels := []string{node.Name}
+	for _, label := range m.InheritedLabels {
+		value, _ := node.Labels.Get(label)
+		commonLabels = append(commonLabels, value)
+	}
+
+	// Remove old metrics before setting the new ones, because otherwise we may end up with
+	// un-updated metrics if node labels change.
+	m.Remove(node)
+
+	for _, f := range node.CPU.Fields() {
+		//nolint:gocritic // assigning append value to a different slice is intentional here
+		labels := append(commonLabels, f.Name)
+		m.cpu.WithLabelValues(labels...).Set(f.Value.AsFloat64())
+	}
+	for _, f := range node.Mem.Fields() {
+		//nolint:gocritic // assigning append value to a different slice is intentional here
+		labels := append(commonLabels, f.Name)
+		m.mem.WithLabelValues(labels...).Set(f.Value.AsFloat64())
+	}
+}
+
+func (m *nodeMetrics) Remove(node *state.Node) {
+	baseMatch := prometheus.Labels{"node": node.Name}
+	m.cpu.DeletePartialMatch(baseMatch)
+	m.mem.DeletePartialMatch(baseMatch)
+}
